+++ CHANGELOG.md
@@ -609,6 +609,7 @@ https://docs.voltocms.com/upgrade-guide/
 
 - Create link in Draftjs using Objectbrowser @giuliaghisini
 
+ - add shouldComponentUpdate to blocks @nileshgulia1
 ### Bugfix
 
 - Allow select widget to reset when the incoming props change. The react-select widget has its own internal state, so if you initialise the widget without choices, then populate the choices, it wouldn't properly show the default value @tiberiuichim